feat: Update zsh configs to support multiple Homebrew installations
feat #1: Update .zshenv to support multiple Homebrew installations

Added conditional checks in .zshenv to evaluate the Homebrew environment
based on the existence of the Homebrew binary in either the /opt/homebrew
or /usr/local/bin directories. This ensures compatibility with different
Homebrew installations.

feat #2: Improve Antigen sourcing logic in .zshrc

Updated the .zshrc configuration to include conditional sourcing for Antigen.
Checks for the existence of Antigen in two possible locations and provides
a message if not found, enhancing user experience and installation guidance.